feat: rename hook
