+++ src/webapp/components/user-role/UserRoleTable.tsx
@@ -4,7 +4,7 @@ import i18n from "../../../utils/i18n";
 import { UserRole } from "../../../domain/entities/UserRole";
 import { ObjectsList, TableColumn, TableConfig, TablePagination, TableSorting } from "@eyeseetea/d2-ui-components";
 import { useAppContext } from "../../contexts/app-context";
-import { useTableWithSelectionCount } from "../../hooks/useTableWithSelectionBanner";
+import { useTableWithSelectionCount } from "../../hooks/useTableWithSelectionCount";
 import { Pager } from "../../../domain/entities/PaginatedResponse";
 import { buildEllipsizedList } from "../user-list-table/UserListTable";
 import { UsersFilters, UsersFiltersProps } from "../users-filter/UsersFilters";
